+++ my-app/src/Pages/JoinPage/Join.js
